add basic dough recipe
this is an all purpose dough that can be used for different variations
of bread. sometimes you might want to add cheese, walnuts, dates
whatever you feel like. the basics remain the same with each dough you
prepare. the purpose of this small article is to explain you the basics.