Setup basic sourdough recipes
Added a few sourdough recipes that good to start. Detailed recipes
should follow. We also need to find a standardized way to organise all
the sourdough recipes, similar like we have for the yeast recipes. Most
of the steps are very similar so it is safe to abstract the process.