Explain how to re-heat the bread a few days later
This happened to me today. I baked on the weekend and brought a full
loaf to a couple of friends. I used steam with lower temperature to
re-heat the full dough. The result was amazing. The crust did not get
any thicker. The bread was warm and almost as good as on the first day.