Add a note that some ingredients should be excluded (#72)
In some recipes we exclude the ingredients. This is because we do not
want the ingredients to influence the dough too much. Instead we add the
ingredients before the final shaping.

In case you want to make sure the ingredients are fully homogenized you
can add them right away initially. However that mightinfluence the way
how you have to handle the dough.