Add Experiments section to the Readme
For some time I wanted to add experiments where only one parameter
changes between each subject of the experiment. I hope more people will
submit some experiments. That way we can all learn much faster and do
not have to repeat all the experiments on our own.

I added an actual experiment and a template for the experiment. Use that
template for the experiments that you want to conduct.