Proofing Sourdough at room temperature vs. fridge
Test what impact proofing a sourdough bread has at room temperature vs. in the
fridge. I long had the feeling that if I bake cold sourdough bread it has more
oven raise than a room temperature proofed one. The process of
letting a dough rest in the fridge is also known as retarding. Because of the
low temperature in the fridge the yeasts' and bacteria's' activity is reduced,
slowing the whole fermentation process.