Emmer sourdough bread
Emmer wheat, is also known as farro especially in Italy, or hulled wheat.
Along with einkorn wheat, emmer was one of the first crops domesticated
in the Near East. It was widely cultivated in the ancient world, but is
now a relict crop in mountainous regions of Europe and Asia.
Its value lies in its ability to give good yields on poor soils,
and its resistance to fungal diseases.