Einkorn bread recipe (#102)
This makes use of the starter to create another great bread.

Einkorn wheat was one of the first plants to be domesticated and cultivated.
The earliest clear evidence of the domestication of einkorn dates from 10,600
to 9,900 years before present. Remnants of einkorn were found with the iceman
mummy Ötzi, dated to 3100 BCE. Einkorn wheat is low-yielding but can survive
 on poor, dry, marginal soils where other varieties of wheat will not.

Einkorn has low gluten and thus will not rise as much in the oven as a
wheat bread would do. Since there is little Gluten there is also no
need to stretch and fold the dough. You mostly mix the ingredients and wait.