Setup Recipe Parser (#106)
Since our templates inherit from each other we need to understand the
content of each of the templates. My long term vision is to create a
user friendly version in which all the templates are properly merged
together. Then you can just read one of the recipes. It will be
automatically added in the proper template.

At the same time I still want markdown to be used for all the recipes.
That makes it very comfortable to write recipes. At the same time it is
also a little easier for everyone to follow the recipes directly in this
git repository.